Crate task_executor

Modules

broadcast

A multi-producer, multi-consumer broadcast queue. Each sent value is seen by all consumers.

signal

Asynchronous signal handling for Tokio.

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

SignalKind

Represents the specific kind of signal to listen for.

Traits

Typesense

Functions

email_executor

generate_trending

main

pulsar_email

pulsar_relation

pulsar_typesense

relation_executor

search_executor

signal

Creates a new stream which will receive notifications when the current process receives the specified signal kind.

trending_executor